+++ src/Argument/Argument.php
@@ -3,6 +3,7 @@
 namespace League\CLImate\Argument;
 
 use League\CLImate\Exceptions\UnexpectedValueException;
+
 use function is_array;
 
 class Argument
@@ -262,8 +263,11 @@ class Argument
      */
     protected function setNoValue($noValue)
     {
-        $this->setCastTo('bool');
         $this->noValue = (bool) $noValue;
+
+        if ($this->noValue === true) {
+            $this->setCastTo('bool');
+        }
     }
 
     /**