+++ src/TerminalObject/Dynamic/Checkbox/Checkbox.php
@@ -8,7 +8,9 @@ use League\CLImate\Util\UtilImporter;
 
 class Checkbox
 {
-    use StringLength, ParserImporter, UtilImporter;
+    use StringLength;
+    use ParserImporter;
+    use UtilImporter;
 
     /**
      * The value of the checkbox