+++ src/TerminalObject/Dynamic/Confirm.php
@@ -3,14 +3,13 @@
 namespace League\CLImate\TerminalObject\Dynamic;
 
 use League\CLImate\Util\Reader\ReaderInterface;
+
 use function in_array;
 use function strtolower;
 use function substr;
 
 class Confirm extends Input
 {
-
-
     /**
      * @inheritdoc
      */