+++ tests/Argument/ManagerTest.php
@@ -12,7 +12,7 @@ class ManagerTest extends TestCase
 
     public function setUp(): void
     {
-        $this->manager = new Manager;
+        $this->manager = new Manager();
     }
 
 
@@ -69,5 +69,37 @@ class ManagerTest extends TestCase
 
         $this->assertEquals('test trailing with spaces', $this->manager->trailing());
         $this->assertEquals(['test', 'trailing with spaces'], $this->manager->trailingArray());
+    }
+
+    public function testItSuggestAlternativesToUnknowArguments()
+    {
+        $this->manager->add([
+            'user' => [
+                'longPrefix' => 'user',
+            ],
+            'password' => [
+                'longPrefix' => 'password',
+            ],
+            'flag' => [
+                'longPrefix' => 'flag',
+                'noValue'    => true,
+            ],
+        ]);
+
+        $argv = [
+            'test-script',
+            '--user=baz',
+            '--pass=123',
+            '--fag',
+            '--xyz',
+        ];
+
+        $this->manager->parse($argv);
+        $processed = $this->manager->getUnknowPrefixedArgumentsAndSuggestions();
+
+        $this->assertCount(3, $processed);
+        $this->assertEquals('password', $processed['pass']);
+        $this->assertEquals('flag', $processed['fag']);
+        $this->assertEquals('', $processed['xyz']);
     }
 }