+++ tests/FileTest.php
@@ -15,6 +15,7 @@ class FileTest extends TestBase
 
     public function setUp(): void
     {
+        parent::setUp();
         $root       = vfsStream::setup();
         $this->file = vfsStream::newFile('log')->at($root);
     }
@@ -34,7 +35,7 @@ class FileTest extends TestBase
                         ->with($this->file->url(), 'a')
                         ->andReturn(fopen($this->file->url(), 'a'));
 
-        $output = new Output;
+        $output = new Output();
         $output->add('file', $file);
         $output->defaultTo('file');
 
@@ -48,7 +49,7 @@ class FileTest extends TestBase
     {
         $resource = fopen($this->file->url(), 'a');
         $file     = $this->getFileMock($resource);
-        $output   = new Output;
+        $output   = new Output();
         $output->add('file', $file);
         $output->defaultTo('file');
 
@@ -73,7 +74,7 @@ class FileTest extends TestBase
 
         $file->lock();
 
-        $output = new Output;
+        $output = new Output();
         $output->add('file', $file);
         $output->defaultTo('file');
 
@@ -119,7 +120,7 @@ class FileTest extends TestBase
         $this->expectExceptionMessage("is not writable");
 
         $file   = $this->getFileMock($this->file->url());
-        $output = new Output;
+        $output = new Output();
         $output->add('file', $file);
         $output->defaultTo('file');
 
@@ -133,7 +134,7 @@ class FileTest extends TestBase
         $this->expectExceptionMessage("The resource [something-that-doesnt-exist] is not writable");
 
         $file   = $this->getFileMock('something-that-doesnt-exist');
-        $output = new Output;
+        $output = new Output();
         $output->add('file', $file);
         $output->defaultTo('file');
 
@@ -153,7 +154,7 @@ class FileTest extends TestBase
                         ->with($this->file->url(), 'a')
                         ->andReturn(false);
 
-        $output = new Output;
+        $output = new Output();
         $output->add('file', $file);
         $output->defaultTo('file');