+++ CHANGELOG.md
@@ -3,6 +3,35 @@ Changelog
 
 --------
 
+## 3.8.2 - 2022-01-23
+
+### Fixed
+
+* Avoid passing null to strlen(). [#186](https://github.com/thephpleague/climate/issues/186)
+* [Support] Added compatibility for psr/log 2 & 3. [#191](https://github.com/thephpleague/climate/issues/191)
+
+--------
+
+## 3.8.1 - 2022-01-23
+
+### Fixed
+
+* [Linux] Avoid type issue when checking if \STDOUT is defined. [#185](https://github.com/thephpleague/climate/issues/185)
+
+--------
+
+## 3.8.0 - 2022-01-22
+
+### Changed
+
+* [Support] Added support for PHP 8.1.
+
+### Fixed
+
+* [Linux] Added a workaround for executed some parts of the code in a non-cli context. [#175](https://github.com/thephpleague/climate/pull/175)
+
+--------
+
 ## 3.7.0 - 2021-01-10
 
 ### Changed